Fix issues and comments
